+++ Repo/user/quake/keys.c
@@ -237,8 +237,8 @@ void Key_Console (int key)
 	if (key == K_PGUP || key==K_MWHEELUP)
 	{
 		con_backscroll += 2;
-		if (con_backscroll > con_totallines - (vid.height>>3) - 1)
-			con_backscroll = con_totallines - (vid.height>>3) - 1;
+		if (con_backscroll > con_totallines - ((int)vid.height>>3) - 1)
+			con_backscroll = con_totallines - ((int)vid.height>>3) - 1;
 		return;
 	}
 
@@ -252,7 +252,7 @@ void Key_Console (int key)
 
 	if (key == K_HOME)
 	{
-		con_backscroll = con_totallines - (vid.height>>3) - 1;
+		con_backscroll = con_totallines - ((int)vid.height>>3) - 1;
 		return;
 	}
 
@@ -756,4 +756,3 @@ void Key_ClearStates (void)
 		key_repeats[i] = 0;
 	}
 }
-
